+++ nodes/knxUltimateHueLight.js
@@ -104,13 +104,24 @@ module.exports = function (RED) {
     // Used to call the status update from the config node.
     node.setNodeStatus = ({
       fill, shape, text, payload,
-    }) => { };
+    }) => {
+      try {
+        if (payload === undefined) payload = '';
+        const dDate = new Date();
+        payload = typeof payload === "object" ? JSON.stringify(payload) : payload.toString();
+        node.sKNXNodeStatusText = `|KNX: ${text} ${payload} (${dDate.getDate()}, ${dDate.toLocaleTimeString()})`;
+        node.status({ fill, shape, text: (node.sHUENodeStatusText || '') + ' ' + (node.sKNXNodeStatusText || '') });
+      } catch (error) { }
+    };
     // Used to call the status update from the HUE config node.
     node.setNodeStatusHue = ({ fill, shape, text, payload }) => {
-      if (payload === undefined) payload = '';
-      const dDate = new Date();
-      payload = typeof payload === "object" ? JSON.stringify(payload) : payload.toString();
-      node.status({ fill, shape, text: `${text} ${payload} (${dDate.getDate()}, ${dDate.toLocaleTimeString()})` });
+      try {
+        if (payload === undefined) payload = '';
+        const dDate = new Date();
+        payload = typeof payload === "object" ? JSON.stringify(payload) : payload.toString();
+        node.sHUENodeStatusText = `|HUE: ${text} ${payload} (${dDate.getDate()}, ${dDate.toLocaleTimeString()})`;
+        node.status({ fill, shape, text: node.sHUENodeStatusText + ' ' + (node.sKNXNodeStatusText || '') });
+      } catch (error) { }
     };
 
     function getRandomIntInclusive(min, max) {